+++ pkg/local_object_storage/shard/control.go
@@ -245,9 +245,8 @@ func (s *Shard) resyncObjectHandler(addr oid.Address, data []byte) error {
 		}
 	}
 
-	err := s.metaBase.Put(obj)
-	if err != nil && !meta.IsErrRemoved(err) && !errors.Is(err, meta.ErrObjectIsExpired) &&
-		!errors.Is(err, apistatus.ErrObjectLocked) {
+	err := s.metaBase.PutResync(obj)
+	if err != nil && !errors.Is(err, meta.ErrObjectIsExpired) {
 		return err
 	}