+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog for NeoFS Node
 - Incorrect garbage bucket setting in metabase on tombstone writing (#3484)
 - Incorrect garbage bucket items via metabase migration (#3484)
 - Rare IR deadlocks during contract deployment (#3237)
+- Metabase resync for objects which are both tombstoned and locked (#3487)
 
 ### Changed
 - CLI uses dynamic polling interval now depending on block time (half a block, but not less than 50ms), this reduces delays for fast networks (#3488)